Added Bar Plots and Scatter Plots
